streamlit_wrt_app/requirements.txt
streamlit>=1.30.0
openpyxl>=3.1.2
python-docx>=1.1.0
requests>=2.31.0
pandas>=2.1.4